refactor(signals): use entrypoints instead of namespace packages
Fixes #4691
