+++ mergify_engine/usage/last_seen.py
@@ -17,12 +17,12 @@ import typing
 
 from mergify_engine import date
 from mergify_engine import github_types
+from mergify_engine import signals
 
 
 if typing.TYPE_CHECKING:
     from mergify_engine import context
     from mergify_engine import redis_utils
-    from mergify_engine import signals
 
 # Keep data 60 days after last signal
 RETENTION_SECONDS = int(datetime.timedelta(days=60).total_seconds())
@@ -32,14 +32,17 @@ def get_last_seen_key(account_id: github_types.GitHubAccountIdType) -> str:
     return f"usage/last_seen/{account_id}"
 
 
-async def update(
-    repository: "context.Repository",
-    event: "signals.EventName",
-    metadata: "signals.EventMetadata",
-) -> None:
-    key = get_last_seen_key(repository.installation.owner_id)
-    now = date.utcnow().isoformat()
-    await repository.installation.redis.cache.setex(key, RETENTION_SECONDS, now)
+class Signal(signals.SignalBase):
+    async def __call__(
+        self,
+        repository: "context.Repository",
+        pull_request: github_types.GitHubPullRequestNumber,
+        event: signals.EventName,
+        metadata: signals.EventMetadata,
+    ) -> None:
+        key = get_last_seen_key(repository.installation.owner_id)
+        now = date.utcnow().isoformat()
+        await repository.installation.redis.cache.setex(key, RETENTION_SECONDS, now)
 
 
 async def get(
