--- mergify_engine_signals/usage/__init__.py
@@ -1,34 +1,0 @@
-# -*- encoding: utf-8 -*-
-#
-# Licensed under the Apache License, Version 2.0 (the "License"); you may
-# not use this file except in compliance with the License. You may obtain
-# a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
-# License for the specific language governing permissions and limitations
-# under the License.
-
-import typing
-
-from mergify_engine import github_types
-from mergify_engine import signals
-from mergify_engine.usage import last_seen
-
-
-if typing.TYPE_CHECKING:
-    from mergify_engine import context
-
-
-class Signal(signals.SignalBase):
-    async def __call__(
-        self,
-        repository: "context.Repository",
-        pull_request: github_types.GitHubPullRequestNumber,
-        event: signals.EventName,
-        metadata: signals.EventMetadata,
-    ) -> None:
-        await last_seen.update(repository, event, metadata)
