+++ setup.cfg
@@ -16,9 +16,6 @@ classifier =
 [options]
 packages =
     mergify_engine
-    mergify_engine_signals.noop
-    mergify_engine_signals.usage
-    mergify_engine_signals._datadog
 
 include_package_data = true
 
@@ -131,6 +128,11 @@ mergify_actions =
     squash = mergify_engine.actions.squash:SquashAction
     edit = mergify_engine.actions.edit:EditAction
 
+mergify_signals =
+    noop = mergify_engine.signals:NoopSignal
+    last_seen = mergify_engine.usage.last_seen:Signal
+    datadog = mergify_engine._datadog:Signal
+
 [build_wheel]
 universal = 1
 
@@ -141,7 +143,7 @@ warn_unused_ignores = true
 warn_unused_configs = true
 disallow_any_generics = true
 warn_return_any = true
-files = mergify_engine,mergify_engine_signals
+files = mergify_engine
 disallow_subclassing_any = true
 warn_redundant_casts = true
 strict_equality = true
